Remove inline semantic layer from dashboards
Dashboards could define a `semantic:` block inline (source/metrics/
dimensions) as an alternative to referencing a separate semantic model
file. The inline form was a parallel, limited implementation: no
segments, no formatting, no window functions, dimensions restricted to a
single column + date type, and it was never documented.

Remove it entirely so the semantic layer has a single source of truth —
the separate-file models in `semantic/*.yml` referenced via `model:`.

- Delete the inline SemanticLayer/Source/Metric/Dimension types and the
  metrics.go SQL/expression generation they fed.
- Drop the merged-metrics fanout machinery (MetricFanout, MetricsJobID,
  Fanout*) and the inline branches in query resolution and validation;
  each widget now resolves to a single job.
- Remove the <Semantic> TSX tag and the inline semantic JSON schema defs.
- metric/metrics/dimension widget fields now require a semantic model
  (set `model:` on the widget or dashboard) instead of silently falling
  back to an inline definition.

The separate-file semantic layer (pkg/semantic) is unchanged.
