--- testdata/dashboards/google-analytics.yml
@@ -1,164 +1,0 @@
-schema: https://getbruin.com/schemas/dac/dashboard/v1
-name: Google Analytics
-description: Web analytics across Bruin properties
-connection: gcp-default
-
-filters:
-  - name: property
-    type: select
-    default: "Bruin Cloud"
-    options:
-      values: ["Bruin Cloud", "Ingestr Docs", "Bruin Docs"]
-
-  - name: date_range
-    type: date-range
-    default:
-      start: "2025-06-01"
-      end: "2026-03-06"
-
-semantic:
-  source:
-    table: "`bruin-landing-page.{% if filters.property == 'Bruin Cloud' %}analytics_417958677{% elif filters.property == 'Ingestr Docs' %}analytics_428962930{% else %}analytics_469869843{% endif %}.events_intraday_*`"
-    date_column: event_date
-    date_format: "%Y%m%d"
-
-  dimensions:
-    daily:
-      column: event_date
-      type: date
-
-    country:
-      column: geo.country
-
-    event:
-      column: event_name
-
-  metrics:
-    page_views:
-      aggregate: count
-      filter:
-        event_name: page_view
-
-    users:
-      aggregate: count_distinct
-      column: user_pseudo_id
-
-    sessions:
-      aggregate: count
-      filter:
-        event_name: session_start
-
-    events:
-      aggregate: count
-
-    pages_per_session:
-      expression: page_views / sessions
-
-rows:
-  - widgets:
-      - name: Page Views
-        type: metric
-        metric: page_views
-        value:
-          field: page_views
-          type: number
-          format: ",.0f"
-        col: 3
-
-      - name: Users
-        type: metric
-        metric: users
-        value:
-          field: users
-          type: number
-          format: ",.0f"
-        col: 3
-
-      - name: Sessions
-        type: metric
-        metric: sessions
-        value:
-          field: sessions
-          type: number
-          format: ",.0f"
-        col: 3
-
-      - name: Pages / Session
-        type: metric
-        metric: pages_per_session
-        col: 3
-
-  - widgets:
-      - name: Daily Traffic
-        type: chart
-        chart: area
-        dimension: daily
-        metrics: [page_views, users]
-        col: 8
-
-      - name: Pages / Session Trend
-        type: chart
-        chart: line
-        dimension: daily
-        metrics: [pages_per_session]
-        col: 4
-
-  - widgets:
-      - name: Traffic Sources
-        type: chart
-        chart: bar
-        col: 4
-        sql: |
-          SELECT
-            IFNULL(traffic_source.source, '(direct)') as source,
-            COUNT(DISTINCT user_pseudo_id) as users
-          FROM `bruin-landing-page.{% if filters.property == 'Bruin Cloud' %}analytics_417958677{% elif filters.property == 'Ingestr Docs' %}analytics_428962930{% else %}analytics_469869843{% endif %}.events_intraday_*`
-          WHERE PARSE_DATE('%Y%m%d', event_date) >= DATE('{{ filters.date_range.start }}')
-            AND PARSE_DATE('%Y%m%d', event_date) <= DATE('{{ filters.date_range.end }}')
-          GROUP BY 1
-          ORDER BY 2 DESC
-          LIMIT 8
-        x: { field: source }
-        y: { field: [users] }
-
-      - name: Top Countries
-        type: chart
-        chart: bar
-        dimension: country
-        metrics: [users]
-        limit: 8
-        col: 4
-
-      - name: Event Breakdown
-        type: chart
-        chart: bar
-        dimension: event
-        metrics: [events]
-        limit: 8
-        col: 4
-
-  - widgets:
-      - name: Top Pages
-        type: table
-        col: 12
-        sql: |
-          SELECT
-            (SELECT value.string_value FROM UNNEST(event_params) WHERE key = 'page_title' LIMIT 1) as page,
-            COUNT(*) as views,
-            COUNT(DISTINCT user_pseudo_id) as users
-          FROM `bruin-landing-page.{% if filters.property == 'Bruin Cloud' %}analytics_417958677{% elif filters.property == 'Ingestr Docs' %}analytics_428962930{% else %}analytics_469869843{% endif %}.events_intraday_*`
-          WHERE event_name = 'page_view'
-            AND PARSE_DATE('%Y%m%d', event_date) >= DATE('{{ filters.date_range.start }}')
-            AND PARSE_DATE('%Y%m%d', event_date) <= DATE('{{ filters.date_range.end }}')
-          GROUP BY 1
-          ORDER BY 2 DESC
-          LIMIT 20
-        columns:
-          - name: page
-            label: Page
-          - name: views
-            label: Views
-            format: number
-          - name: users
-            label: Users
-            format: number
